Add Feasibility / Thermo filters and refactor
